Fix collaborative-editor e2e test expectations and consolidate RGA tests
- Fix expected length in concurrent write test (9 chars instead of 6)
  * 'invitees' target executes on BOTH invitee nodes (nodes 2 & 3)
  * Result: ABC (inviter) + XYZ (invitee 1) + XYZ (invitee 2) = 9 chars
- Update test descriptions to clarify 3-way concurrent write scenario
- Merge rga_sync.rs serialization tests into rga.rs for better organization
- Remove separate rga_sync test module

All 34 collaborative-editor test steps now pass with full CRDT convergence verified.
