+++ e2e-tests/config/protocols/near/collaborative-editor-test.json
@@ -314,7 +314,7 @@
         "expectedResultJson": null,
         "target": "invitees",
         "description": [
-          "Concurrent write test: Node 2 inserts XYZ at position 0 (no wait = concurrent)"
+          "Concurrent write test: Both invitees (nodes 2 & 3) insert XYZ at position 0 simultaneously"
         ]
       }
     },
@@ -324,7 +324,7 @@
         "durationMs": 5000,
         "description": [
           "Wait for concurrent edits to propagate and CRDT to converge.",
-          "Both nodes will converge to same state thanks to RGA CRDT."
+          "All 3 nodes (inviter + 2 invitees) will converge to same state thanks to RGA CRDT."
         ]
       }
     },
@@ -332,12 +332,13 @@
       "call": {
         "methodName": "get_length",
         "argsJson": {},
-        "expectedResultJson": 6,
+        "expectedResultJson": 9,
         "target": "allMembers",
         "retries": 5,
         "intervalMs": 1000,
         "description": [
-          "Verify total length is 6 characters (ABC + XYZ) on all nodes"
+          "Verify total length is 9 characters (ABC + XYZ + XYZ) on all nodes",
+          "Note: 'invitees' target means BOTH invitee nodes insert XYZ concurrently"
         ]
       }
     },
@@ -350,9 +351,9 @@
         "retries": 10,
         "intervalMs": 1000,
         "description": [
-          "CRDT Convergence Test: Verify all nodes have identical text after concurrent edits.",
+          "CRDT Convergence Test: Verify all 3 nodes have identical text after concurrent edits.",
           "The RGA CRDT ensures deterministic ordering based on HLC timestamps.",
-          "Both ABC and XYZ are present, ordered deterministically by their timestamps.",
+          "ABC and two XYZ insertions are present, ordered deterministically by their timestamps.",
           "Using 'allMembers' target ensures the same value is returned from all nodes,",
           "proving CRDT convergence - a key property of conflict-free replicated data types."
         ]
